가게 상세화면 별점 셀 구현

+++ Modules/Common/Targets/Model/Sources/Domain/Response/StoreDetailApiResponse.swift
@@ -4,7 +4,7 @@ public struct StoreDetailApiResponse: Decodable {
     public let storeId: Int
     public let name: String
     public let salesType: String?
-    public let rating: Int
+    public let rating: Double
     public let location: LocationResponse
     public let address: AddressResponse
     public let categories: [PlatformStoreFoodCategoryResponse]
